requirements.txt
arrow==1.2.3
attrs==22.2.0
certifi==2022.12.7
charset-normalizer==2.1.1
icalendar==5.0.3
idna==3.4
PyJWT==2.6.0
python-dateutil==2.8.2
python-dotenv==0.21.0
pytz==2022.7
requests==2.28.1
six==1.16.0
TatSu==5.8.3
twilio==7.16.0
urllib3==1.26.13
